feat(collections): stabilize index argument in iterable methods
