+++ collections/deno.json
@@ -48,23 +48,7 @@
     "./union": "./union.ts",
     "./unstable-binary-search": "./unstable_binary_search.ts",
     "./unstable-cycle": "./unstable_cycle.ts",
-    "./unstable-distinct-by": "./unstable_distinct_by.ts",
-    "./unstable-drop-last-while": "./unstable_drop_last_while.ts",
-    "./unstable-drop-while": "./unstable_drop_while.ts",
-    "./unstable-find-single": "./unstable_find_single.ts",
-    "./unstable-first-not-nullish-of": "./unstable_first_not_nullish_of.ts",
     "./unstable-interleave": "./unstable_interleave.ts",
-    "./unstable-join-to-string": "./unstable_join_to_string.ts",
-    "./unstable-map-not-nullish": "./unstable_map_not_nullish.ts",
-    "./unstable-max-by": "./unstable_max_by.ts",
-    "./unstable-max-of": "./unstable_max_of.ts",
-    "./unstable-min-by": "./unstable_min_by.ts",
-    "./unstable-min-of": "./unstable_min_of.ts",
-    "./unstable-partition": "./unstable_partition.ts",
-    "./unstable-sort-by": "./unstable_sort_by.ts",
-    "./unstable-sum-of": "./unstable_sum_of.ts",
-    "./unstable-take-last-while": "./unstable_take_last_while.ts",
-    "./unstable-take-while": "./unstable_take_while.ts",
     "./unzip": "./unzip.ts",
     "./without-all": "./without_all.ts",
     "./zip": "./zip.ts"
