+++ collections/drop_while.ts
@@ -8,7 +8,8 @@
  * @typeParam T The type of the elements in the input iterable.
  *
  * @param iterable The iterable to drop elements from.
- * @param predicate The function to test each element for a condition.
+ * @param predicate The function to test each element for a condition. The
+ * function receives the element and its index.
  *
  * @returns An array that drops all elements until the first element that
  * does not match the given predicate.
@@ -23,22 +24,34 @@
  *
  * assertEquals(dropWhileNumbers, [2, 5, 2, 5]);
  * ```
+ *
+ * @example Using the index parameter
+ * ```ts
+ * import { dropWhile } from "@std/collections/drop-while";
+ * import { assertEquals } from "@std/assert";
+ *
+ * const array = [20, 30, 20];
+ * const result = dropWhile(array, (_, index) => index < 1);
+ *
+ * assertEquals(result, [30, 20]);
+ * ```
  */
 export function dropWhile<T>(
   iterable: Iterable<T>,
-  predicate: (el: T) => boolean,
+  predicate: (el: T, index: number) => boolean,
 ): T[] {
   if (Array.isArray(iterable)) {
-    const idx = iterable.findIndex((el) => !predicate(el));
+    const idx = iterable.findIndex((el, index) => !predicate(el, index));
     if (idx === -1) {
       return [];
     }
     return iterable.slice(idx);
   }
   const array: T[] = [];
+  let index = 0;
   let found = false;
   for (const item of iterable) {
-    if (found || !predicate(item)) {
+    if (found || !predicate(item, index++)) {
       found = true;
       array.push(item);
     }
