+++ collections/find_single.ts
@@ -8,7 +8,8 @@
  * @typeParam T The type of the elements in the input array.
  *
  * @param array The array to find a single element in.
- * @param predicate The function to test each element for a condition.
+ * @param predicate The function to test each element for a condition. The
+ * function receives the element and its index.
  *
  * @returns The single element that matches the given condition or `undefined`
  * if there are zero or more than one matching elements.
@@ -29,15 +30,27 @@
  * assertEquals(activeBooking, { month: "June", active: true });
  * assertEquals(inactiveBooking, undefined); // There are two applicable items
  * ```
+ *
+ * @example Using the index parameter
+ * ```ts
+ * import { findSingle } from "@std/collections/find-single";
+ * import { assertEquals } from "@std/assert";
+ *
+ * const array = [9, 12, 13];
+ * const result = findSingle(array, (_, index) => index === 1);
+ *
+ * assertEquals(result, 12);
+ * ```
  */
 export function findSingle<T>(
   array: Iterable<T>,
-  predicate: (el: T) => boolean,
+  predicate: (el: T, index: number) => boolean,
 ): T | undefined {
   let match: T | undefined;
   let found = false;
+  let index = 0;
   for (const element of array) {
-    if (predicate(element)) {
+    if (predicate(element, index++)) {
       if (found) return undefined;
       found = true;
       match = element;
