+++ collections/first_not_nullish_of_test.ts
@@ -2,7 +2,6 @@
 
 import { assertEquals } from "@std/assert";
 import { firstNotNullishOf } from "./first_not_nullish_of.ts";
-import * as unstable from "./unstable_first_not_nullish_of.ts";
 
 function firstNotNullishOfTest<T, O>(
   input: [Array<T>, (el: T) => O | undefined | null],
@@ -88,9 +87,9 @@ Deno.test({
 });
 
 Deno.test({
-  name: "unstable.firstNotNullishOf() passes index to selector",
+  name: "firstNotNullishOf() passes index to selector",
   fn() {
-    const result = unstable.firstNotNullishOf(
+    const result = firstNotNullishOf(
       [1, 2, 3, 4],
       (it, index) => index < 1 ? null : it,
     );
