+++ collections/join_to_string_test.ts
@@ -2,7 +2,6 @@
 
 import { assertEquals } from "@std/assert";
 import { joinToString } from "./join_to_string.ts";
-import * as unstable from "./unstable_join_to_string.ts";
 
 Deno.test({
   name: "joinToString() handles no mutation",
@@ -147,11 +146,11 @@ Deno.test({
 });
 
 Deno.test({
-  name: "unstable.joinToString() passes index to selector",
+  name: "joinToString() passes index to selector",
   fn() {
     const arr = ["Kim", "Anna", "Tim"];
 
-    const out = unstable.joinToString(arr, (it, index) => it + index);
+    const out = joinToString(arr, (it, index) => it + index);
 
     assertEquals(out, "Kim0,Anna1,Tim2");
   },
