+++ collections/max_of.ts
@@ -32,7 +32,7 @@
  */
 export function maxOf<T>(
   array: Iterable<T>,
-  selector: (el: T) => number,
+  selector: (el: T, index: number) => number,
 ): number | undefined;
 /**
  * Applies the given selector to all elements of the provided collection and
@@ -65,9 +65,14 @@ export function maxOf<T>(
  */
 export function maxOf<T>(
   array: Iterable<T>,
-  selector: (el: T) => bigint,
+  selector: (el: T, index: number) => bigint,
 ): bigint | undefined;
-export function maxOf<T, S extends ((el: T) => number) | ((el: T) => bigint)>(
+export function maxOf<
+  T,
+  S extends
+    | ((el: T, index: number) => number)
+    | ((el: T, index: number) => bigint),
+>(
   array: Iterable<T>,
   selector: S,
 ): ReturnType<S> | undefined {
@@ -75,11 +80,11 @@ export function maxOf<T, S extends ((el: T) => number) | ((el: T) => bigint)>(
     const length = array.length;
     if (length === 0) return undefined;
 
-    let max = selector(array[0]!) as ReturnType<S>;
+    let max = selector(array[0]!, 0) as ReturnType<S>;
     if (Number.isNaN(max)) return max;
 
     for (let i = 1; i < length; i++) {
-      const currentValue = selector(array[i]!) as ReturnType<S>;
+      const currentValue = selector(array[i]!, i) as ReturnType<S>;
       if (currentValue > max) {
         max = currentValue;
       } else if (Number.isNaN(currentValue)) {
@@ -90,17 +95,18 @@ export function maxOf<T, S extends ((el: T) => number) | ((el: T) => bigint)>(
     return max;
   }
 
+  let index = 0;
   const iter = array[Symbol.iterator]();
   const first = iter.next();
 
   if (first.done) return undefined;
 
-  let max = selector(first.value) as ReturnType<S>;
+  let max = selector(first.value, index++) as ReturnType<S>;
   if (Number.isNaN(max)) return max;
 
   let next = iter.next();
   while (!next.done) {
-    const currentValue = selector(next.value) as ReturnType<S>;
+    const currentValue = selector(next.value, index++) as ReturnType<S>;
     if (currentValue > max) {
       max = currentValue;
     } else if (Number.isNaN(currentValue)) {
