+++ collections/min_of_test.ts
@@ -1,7 +1,6 @@
 // Copyright 2018-2026 the Deno authors. MIT license.
 import { minOf } from "./min_of.ts";
 import { assertEquals } from "@std/assert";
-import * as unstable from "./unstable_min_of.ts";
 
 Deno.test("minOf() handles regular min", () => {
   const array = [5, 18, 35, 120];
@@ -145,11 +144,11 @@ Deno.test("minOf() handles minus infinity", () => {
 });
 
 Deno.test({
-  name: "unstable.minOf() passes index to selector",
+  name: "minOf() passes index to selector",
   fn() {
     const input = [4, 3, 2, 1];
 
-    const max = unstable.minOf(input, (it, index) => it * index);
+    const max = minOf(input, (it, index) => it * index);
 
     assertEquals(max, 0);
   },
