+++ collections/partition_test.ts
@@ -2,7 +2,6 @@
 
 import { assertEquals } from "@std/assert";
 import { partition } from "./partition.ts";
-import * as unstable from "./unstable_partition.ts";
 
 function partitionTest<I>(
   input: [Array<I>, (element: I) => boolean],
@@ -115,9 +114,9 @@ Deno.test({
 });
 
 Deno.test({
-  name: "unstable.partition() passes index to predicate",
+  name: "partition() passes index to predicate",
   fn() {
-    const result = unstable.partition(
+    const result = partition(
       [2, 4, 6],
       (_: number, index: number) => index % 2 === 0,
     );
