+++ collections/sort_by_test.ts
@@ -2,7 +2,6 @@
 
 import { assertEquals } from "@std/assert";
 import { sortBy } from "./sort_by.ts";
-import * as unstable from "./unstable_sort_by.ts";
 
 Deno.test({
   name: "sortBy() handles no mutation",
@@ -334,8 +333,8 @@ Deno.test({
 });
 
 Deno.test({
-  name: "unstable.sortBy() passes index to selector",
+  name: "sortBy() passes index to selector",
   fn() {
-    assertEquals(unstable.sortBy([2, 3, 1], (_, index) => -index), [1, 3, 2]);
+    assertEquals(sortBy([2, 3, 1], (_, index) => -index), [1, 3, 2]);
   },
 });
